Revised initial version
So, some quick debugging and a change of name "asm6809" was used by so many people as it is the obvious choice.  "asm68c9" seems to be acceptably unique.
Testing has fixed a range of problems from the earlier version, and this version now seems to with some accuracy.  I say this as it has successfully assembled the Motorola ASSIST09 source code and (as far as I have managed to verify so far) generated the same target machine code and data.