Functionally extended
Actually more useful now with output options for Intel Hex and Motorola S records ins addition to just plain hexadecimal data.  Actual testing of the assembler is still subject to real world use .. this is dependent on me actually making a computer based on a 6809 (or writing that emulator I had planned).