SETDP issue identified and resolved
Fixed (hacked) issues with SETDP being assigned a 16 bit value.  Took pragmatic approach rather than simply producing an error.  I have included the "assist09" source code I have access to.  This seems to be the original Motorola version and assembles using the legacy option.  Included the 'hex' and 'srec' output for "assist09" as produced by this assembler.
NB/  I have had to fiddle just one line: the SETDP directive.  It was initially commented out and the target label preceded by a '=' operator.  The solution (for the moment) is removing the comment character and unrecognised operator and applying the 16-bit SETDP hack.